M-x sly reuse disconnected existing inferior-lisp buffers
sly.el: (sly--lookup-lisp-implementation): reuse existing
inferior-lisp buffers with M-x sly starting a lisp from
sly-lisp-implementations. The `name' in the alist is assumed to be
identical to (connection-name) which is assigned after the connection
is established.
(maybe-start-lisp): Only prompt for opening an additional
inferior-lisp if there is no active connection or if the connection is
in use. (this differs from earlier behavior)
(sly): make (sly "ecl") start a program ecl as documented.
(sly-set-connection-info): Do not tramplingly rename inferior lisp
buffers