adapt sly-mrepl--open-dedicated-stream for sly-secret upstream changes
The upstream change was retarded. It assumes that the dedicated output
stream is always a binary stream. This assumption is wrong. Clisp for
eg. creates only a character stream for the dedicated output stream
when an external-format is provided. SO to support sly-secret with the
upstream change we have to read a packet from a character
stream.

* contrib/sly-mrepl.el: (sly-mrepl--open-dedicated-stream): adapt to
sly-secret upstream change. sly-secret is not a string but an sexp
transmitted and received via a rpc packet.

* slynk/sly-rpc.lisp: (read-packet-character-stream):
export. (read-chunk-character-stream, parse-header-character,
read-packet-character): new. (authenticate-client): integrate the new
functions to read a packet if the client is a character-stream.