* lib/sly-completion.el: (sly-read-from-minibuffer): tweak minibuffer setup hooks
Ensure that the functions in sly-minibuffer-setup-hook run after
the functions in minibuffer-setup-hook.

The hook functions in sly-minibuffer-setup-hook should be able to set
up completion-at-point-functions for sly-read-from-minibuffer. Eg.

	(add-hook 'sly-minibuffer-setup-hook 'sly-c-p-c-on)

However sly-read-from-minibuffer calls sly--with-sly-minibuffer which
in turn calls sly--setup-completion via minibuffer-completion-hook.
If the functions in sly-minibuffer-setup-hook run before the functions
in minibuffer-completion-hook, then any modifications to
completion-at-point-functions are undone by sly--setup-completion and
we lose.

Tweak the order so the user defined functions run after the functions
that that sly sets up.