turn off sly-autodoc chatter when the module is disabled
* contrib/sly-autodoc.el: (sly-autodoc): use new
eldoc-documentation-functions.  You can remove sly-autodoc from
sly-contribs, unload it, but it will still keep polluting *sly-events*
it is enabled by the broken-by-design of eldoc and minor-mode and
mechanisms. so suicide at startup.

* contrib/sly-mrepl.el: make sly-mrepl contrib not depend on sly-autodoc
so it doesn't pulled in by it. It doesn't really depend on it - the
related variable it binds is bogus.