sly-background-output contrib WIP
*   contrib/sly-background-output.el:
*   contrib/slynk-background-output.lisp:

Send output sent to standard output streams to an emacs buffer called
Background Output rather than to the inferior lisp (when there isn't any
slynk-mrepl)

Work with slynk-mrepls by making swank-mrepl::default-listener a nop
when Background Output is active. Note that this assumes
slynk:*globally-redirect-io* is nil.  You can still kill the last repl
and get output to to Background Output.  M-x sly-background-output-start
sly-background-output-stop to recover in a pinch.

-   ansi-color hooks for background-output

- sly-define-channel-method background-output - inhibit-read-only t so
you can C-x C-q the background output buffer and use q to dismiss it

- 21-07-26
* sly-background-output.el: (sly-background-autofocus): don't raise
background output windows unless this is non-NIL

- 21-09-12
* sly.el: (sly-switch-to-background-output): new
function. (sly-selector-map) add a sly-selector-map binding to `o'.

- 22-02-17 - fix autofocus and debug

- 25-01-06 - add a comment about adding an element to
display-buffer-alist to always hide the background-output-buffer. this
should be integrated with autofocus logic TODO

- 25-01-29
* contrib/sly-background-output.el: (write-string): set-window-start
to top if sly-background-output-scroll-to-bottom is `top', get rid of
warn. set-window-point to point-max if scroll-to-bottom.