sly.el: fix display-buffer argument to sly-with-popup-buffer
* sly.el: (sly-with-popup-buffer) don't pass a non-nil non-list action
argument to display-buffer as that always raises the frame and ignores
the select argument. Instead use alist parameters inhibit-same-window
and inhibit-switch-frame correctly.

If sly-description-autofocus is nil then the sly-description buffer
must not be selected and raised. Without this fix the frame would
always be raised.

[230208]
(sly-with-popup-buffer): always select-frame-set-input-focus on
:select, or M-x sly-repl doesn't raise a different obscured
frame (even with :raise)

[250514]
check window-live-p and window-valid-p (apparently required
on tty)

[250511]
(sly-with-popup-buffer): reuse frames This tries to reuse frames
displaying buffers like *sly-description* (the goal is to reuse them
even if they are invisible with frame-auto-hide-function
 #'make-frame-invisible, and support
`switch-to-buffer-obey-display-actions' t when the user has customized
display-buffer.)

[250529]
keep - always select-frame-set-input-focus, fix parens syntax errors