sly.el: fix display-buffer argument to sly-with-popup-buffer
* sly.el: (sly-with-popup-buffer) don't pass a non-nil non-list action
argument to display-buffer as that always raises the frame and ignores
the select argument. Instead use alist parameters inhibit-same-window
and inhibit-switch-frame correctly.

If sly-description-autofocus is nil then the sly-description buffer
must not be selected and raised. Without this fix the frame would
always be raised.

;madhu 230208
(sly-with-popup-buffer): always select-frame-set-input-focus on
:select, or M-x sly-repl doesn't raise a different obscured
frame (even with :raise)