abcl - defpackage fixes
slynk/backend/abcl.lisp uses slynk and slynk-match packages which it
should not. . it is not possible to replace the use of slynk symbols
by doing a (funcall (find-symbol symbol slynk)) because some of the
uses are of macros.

slime fixed this problem by moving package definitions to a separate
packages.lisp file which is loaded upfront.

in this patch we rearrange the loading order in slynk-loader and
slynk.system so backend/abcl is loaded AFTER slynk-backend and slynk
are loaded.

* slynk/{slynk-loader.lisp, slynk.system}: rearrange file loading
order to satisfy backend/abcl dependencies on
slynk/slynk-backend/slynk-match.

* slynk/slynk-loader.lisp: reperform the use-package and export
functions of defpackage.

typical usage defines a slynk-loader package and sets the
*source-pathname* variable before loading slynk-loader.lisp so that
variable is bound before the defvar so defvar doesn't touch it.

abcl wont process the defpackage form if the package is already
defined so redo what the defpackage form does.

* slynk/slynk-backend.lisp: export gray-package-name and
preferred-communication-style. add a slynk/backend nickname to support
to backend/slynk-abcl.lisp's use of satanic asdf enforced slashes in
package names.

* slynk/slynk-match.lisp: add a slynk/match nickname to support
backend/slynk-abcl.lisp's use of satanic asdf enforced slashes in
package names.