sly.el: fix inspector asynchronicity
* sly.el: (sly--open-inspector):  save-excursion here doesn't work
because the results come asynchronously and the point goes to the
bottom. What is needed is to restore the point in a continuation which
can run after the slime-inspector-insert-content finishes its job.
(sly-inspector-insert-content): signature changed to include an
optional restore-point-cont argument which we call when the chunks are
inserted.

btw sly-insert-chunk-content is  probably an example of where fp
is not the most appropriate