sly-mk-defsystem: jump-to-fasl wip
* contrib/sly-mk-defsystem.el: (jump-to-fasl): new function. dired the
fasl file name. useful for inspecting and deleting stale fasls.

* contrib/slynk-mk-defsystem.lisp: (jump-to-component-1): given a
system and a filename return the component if found.

present assumptions: 1. no complications (no :subsystem or :system
components), 2. the buffer-file name of the visited file is the same
as the full source pathname of the component. this needs
improvement. 3. the filenamestring must be present as a file component
in the either the given defsystem system or in one of the systems in
the depends-on of the given defsystem, provided that
dependent-system's root is within the given defsystem root.