sly.el: (sly-popup-buffer-mode): keymap bind l(back), f(forward) q(quit)
- also switch the define-minor-mode defiition form to "non-obsolete"
form, i.e the first three arguments init-value lighter and keymap
arguments are not positional but keyword arguments now

[251209] - deal with define-minor-mode semantics. the mode documents
that supplying a non-symbol to :keymap arg will define the mode-MAP
var and set it. But if it is already defined this cannot work.

so bind the l and f keys when sly-popup-buffer-mode-map is being
defvar'd.