contrib/sly-mrepl.el: (sly-mrepl-start-in-background): new. don't switch to mrepl buffer on M-x sly
[251022]
* contrib/sly-mrepl.el: (sly-mrepl--goto-connection): new function,
used as the value of sly-connection-list-button-action, use this
instead of the closure. bind sly-mrepl-start-in-background to nil,
because we want to switch.

[260329]

contrib/sly-mrepl.el: move definition of defvar
sly-mrepl-start-in-background to the top of the file, before its use
in sly-mrepl--goto-connection. otherwise (native) compilation captures
binds it as a lexical value